% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{is.adam}
\alias{is.adam}
\alias{is.adam.sim}
\title{ADAM classes checkers}
\usage{
is.adam(x)

is.adam.sim(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if this is the specified class and \code{FALSE} otherwise.
}
\description{
Functions to check if an object is of the specified class
}
\details{
The list of functions includes:
\itemize{
\item \code{is.adam()} tests if the object was produced by a \link[adam]{adam} function
\item \code{is.adam.sim()} tests if the object was produced by \link[adam]{sim.adam} function;
}
}
\examples{

\dontrun{ourModel <- msarima(rnorm(100,100,10))}

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{ts}
\keyword{univar}
