% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omes.R
\name{omes}
\alias{omes}
\title{Occurrence part of the Mixed Exponential Smoothing}
\usage{
omes(...)
}
\arguments{
\item{...}{The parameters of the function}
}
\description{
Function constructs an advanced Single Source of Error model, based on ETS
taxonomy for the probability of occurrence.
}
