% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omes.R
\name{sim.mes}
\alias{sim.mes}
\title{Simulation of the data from the Mixed Exponential Smoothing}
\usage{
sim.mes(...)
}
\arguments{
\item{...}{The parameters of the function}
}
\description{
Function generates the data based on the MES model
}
