% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simomes.R
\name{sim.omes}
\alias{sim.omes}
\title{Simulation of the data from the occurrence part of the Mixed Exponential Smoothing}
\usage{
sim.omes(...)
}
\arguments{
\item{...}{The parameters of the function}
}
\description{
Function generates the data based on the OMES model
}
