% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFunctions.R
\name{is.mes}
\alias{is.mes}
\alias{is.mes.sim}
\title{MES classes checkers}
\usage{
is.mes(x)

is.mes.sim(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if this is the specified class and \code{FALSE} otherwise.
}
\description{
Functions to check if an object is of the specified class
}
\details{
The list of functions includes:
\itemize{
\item \code{is.mes()} tests if the object was produced by a \link[mes]{mes} function
\item \code{is.omes()} tests if the object was produced by the \link[mes]{omes} function;
\item \code{is.mes.sim()} tests if the object was produced by \link[mes]{sim.mes} function;
\item \code{is.omes.sim()} tests if the object was produced by \link[mes]{sim.omes} function;
}
}
\examples{

\dontrun{ourModel <- msarima(rnorm(100,100,10))}

}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{ts}
\keyword{univar}
